+++ inc/pthread_algorithm.h
@@ -1,6 +1,6 @@
 #ifndef PTHREAD_ALGORITHM_H
 #define PTHREAD_ALGORITHM_H
 
-double **calc_heat_plate_pthread(int height, int width, double epsilon);
+float **calc_heat_plate_pthread(int height, int width, float epsilon);
 
 #endif
