+++ inc/serial_algorithm.h
@@ -1,6 +1,6 @@
 #ifndef SERIAL_ALGORITHM_H
 #define SERIAL_ALGORITHM_H
 
-double **calc_heat_plate_serial(int height, int width, double epsilon);
+float **calc_heat_plate_serial(int height, int width, float epsilon);
 
 #endif
