+++ inc/visualization.h
@@ -3,7 +3,7 @@
 
 #include <cv.h>
 
-IplImage *get_image(double **plate, int h, int w);
+IplImage *get_image(float **plate, int h, int w);
 void show_image(IplImage *img);
 void save_image(IplImage *img, const char *file_name);
 void release_image(IplImage *img);
